requirements.txt
opencv-python==4.8.1
numpy==1.26.3
paho-mqtt==2.2.1
firebase-admin==6.2.0
Pillow==10.1.0 